#186: Refactor Statistics assertion classes.
