--- src/test/java/com/commercetools/sync/commons/asserts/syncStatistic/ProductSyncStatisticAssert.java
@@ -1,16 +1,0 @@
-package com.commercetools.sync.commons.asserts.syncStatistic;
-
-import com.commercetools.sync.products.helpers.ProductSyncStatistics;
-
-public class ProductSyncStatisticAssert
-    extends AbstractSyncStatisticAssert<ProductSyncStatisticAssert, ProductSyncStatistics> {
-
-    public ProductSyncStatisticAssert(final ProductSyncStatistics actual) {
-        super(actual, ProductSyncStatisticAssert.class);
-    }
-
-    public static ProductSyncStatisticAssert assertThatStatistic(final ProductSyncStatistics actual) {
-        return new ProductSyncStatisticAssert(actual);
-    }
-
-}
